continued work on primitive type representations.
